requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==2.0.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.3.1
jupyter_core==5.3.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
tzdata==2023.3
wcwidth==0.2.8
